🔧 모달 컴포넌트 수정, 파일 구조 수정, 배경 클릭 및 ESC 키 처리에 대한 props추가
