+++ components/Modal/Modal.tsx
@@ -6,6 +6,8 @@ interface ModalProps {
   onClose: () => void; // 모달 닫기 함수
   children: React.ReactNode; // 모달 내용
   width?: string; // 모달 너비
+  closeOnBackgroundClick?: boolean; // 배경 클릭시 모달 닫기 여부
+  closeOnEsc?: boolean; // esc 키로 모달 닫기 여부
 }
 
 /**
@@ -14,6 +16,8 @@ interface ModalProps {
  * @param onClose 모달 닫기 함수
  * @param children 모달 내용
  * @param width 모달 너비
+ * @param closeOnBackgroundClick 배경 클릭시 닫기 여부 (기본값: false)
+ * @param closeOnEsc ESC 키로 닫기 여부 (기본값: true)
  * @returns Modal 컴포넌트
  */
 
@@ -21,14 +25,16 @@ const Modal = ({
   isOpen,
   onClose,
   children,
-  width = 'w-11/12 ta:w-3/4 pc:w-1/2',
+  width = 'w-3/4 mo:w-11/12 pc:w-[395px]',
+  closeOnBackgroundClick = false,
+  closeOnEsc = true,
 }: ModalProps) => {
   // 모달이 닫혀있으면 null 반환
   if (!isOpen) return null;
 
   // 배경 클릭시 모달 닫기
   const handleBackGroundClick = (e: React.MouseEvent) => {
-    if (e.target === e.currentTarget) {
+    if (closeOnBackgroundClick && e.target === e.currentTarget) {
       onClose();
     }
   };
@@ -36,13 +42,16 @@ const Modal = ({
   // esc 키로 모달 닫기
   useEffect(() => {
     const handleEsc = (e: KeyboardEvent) => {
-      if (e.key === 'Escape') {
+      if (closeOnEsc && e.key === 'Escape') {
         onClose();
       }
     };
-    window.addEventListener('keydown', handleEsc);
-    return () => window.removeEventListener('keydown', handleEsc);
-  }, []);
+
+    if (closeOnEsc) {
+      window.addEventListener('keydown', handleEsc);
+      return () => window.removeEventListener('keydown', handleEsc);
+    }
+  }, [closeOnEsc]);
 
   return (
     <div
@@ -65,7 +74,7 @@ const Modal = ({
           />
         </div>
         {/* 컨텐츠 영역 */}
-        <div className="p-6">{children}</div>
+        <div className="px-5 pb-6 pt-16">{children}</div>
       </div>
     </div>
   );
